Graph ql done

+++ Server/graphql/resolvers/booking.js
@@ -3,8 +3,11 @@ const Booking = require("../../models/booking");
 const { transformBooking, transformEvent } = require("./merge");
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error("Unauthenticated");
+      }
       const bookings = await Booking.find();
       return bookings.map(booking => {
         return transformBooking(booking);
@@ -13,17 +16,23 @@ module.exports = {
       throw err;
     }
   },
-  bookEvent: async args => {
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unauthenticated");
+    }
     const fetchedEvent = await Event.findOne({ _id: args.eventId });
     const booking = new Booking({
-      user: "5e73b108a7933d2acc2b3cee",
+      user: req.userId,
       event: fetchedEvent
     });
     const result = await booking.save();
     return transformBooking(result);
   },
-  cancelBooking: async args => {
+  cancelBooking: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error("Unauthenticated");
+      }
       const booking = await Booking.findById(args.bookingId).populate("event");
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
